컨트롤러 일부 수정(comm->trend), 통합검색 오류 수정
